fix compilation issue for ic_udpifc.c

+++ src/backend/utils/mmgr/memprot.c
@@ -76,7 +76,7 @@ volatile OOMTimeType oomTrackerStartTime = 0;
 #ifndef WIN32
 static pthread_t memprotOwnerThread = (pthread_t)0;
 #else
-static pthread_once_t memprotOwnerThread = (ULONG)0;
+static pthread_once_t memprotOwnerThread = 0;
 #endif
 /* Is memory protection enabled? */
 bool gp_mp_inited = false;
